AC:自動打包到dist目錄

+++ sh/install-js-obfuscator.sh
@@ -1,8 +1,19 @@
 #!/bin/bash
 # for ubuntu/debia/kali
+# initial folder = ~/$PRJ_DIR_NAME/sh
 
+source .env
+
+cd ~
 sudo apt install npm
 sudo npm install --save-dev javascript-obfuscator
-sudo ln -s ~/node_modules/javascript-obfuscator/bin/javascript-obfuscator /usr/local/bin
+if [[ ! -f /usr/local/bin/javascript-obfuscator ]]; then
+    sudo ln -s ~/node_modules/javascript-obfuscator/bin/javascript-obfuscator /usr/local/bin
+fi
 
 echo "Excample: javascript-obfuscator 檔案名稱.js"
+
+source ~/.bashrc
+
+cd ~/$PRJ_DIR_NAME/sh
+
